start in project

+++ get_next_line.h
@@ -6,23 +6,23 @@
 /*   By: ybel-hac <ybel-hac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/29 19:18:11 by ybel-hac          #+#    #+#             */
-/*   Updated: 2022/10/30 19:37:34 by ybel-hac         ###   ########.fr       */
+/*   Updated: 2022/10/31 18:07:36 by ybel-hac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef GET_NEXT_LINE_H
-# define GET_NEXT_LINE_H
+#define GET_NEXT_LINE_H
 
 #include "stdio.h"
 #include <unistd.h>
 #include <fcntl.h>
 #include <stdlib.h>
 
-# ifndef BUFFER_SIZE
-#  define BUFFER_SIZE 4096
-# endif
+#ifndef BUFFER_SIZE
+#define BUFFER_SIZE 4096
+#endif
 
-char	*get_next_line(int fd);
-char	*get_last_str(char *txt, int *current, int *last);
+char *get_next_line(size_t fd);
+char *get_last_str(char *txt, size_t *current, size_t *last);
 
 #endif
